Fix Duplicate product barcode detection

+++ core/domain/src/main/java/com/wansenai/entities/product/ProductExtendPrice.java
@@ -59,7 +59,7 @@ public class ProductExtendPrice implements Serializable {
     /**
      * 商品条码
      */
-    private Integer productBarCode;
+    private Long productBarCode;
 
     /**
      * 商品单位
